requirements.txt
django==3.2.16
Pillow 
pylint
pylint-django
django-bower
xlwt
xlrd
pandas
django-mathfilters
djangorestframework==3.14.0
django-manifest-loader
mssql-django
knox
django-rest-knox
django-pyodbc-azure
boto3
psycopg2
